move views to their own namespace
