+++ GemiNaut/ExternalNavigator.cs
@@ -19,6 +19,8 @@
 //    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //===================================================
 
+using GemiNaut.Views;
+
 namespace GemiNaut
 {
     public class ExternalNavigator
