+++ GemiNaut/Views/Bookmarks.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="GemiNaut.Bookmarks"
+﻿<Window x:Class="GemiNaut.Views.Bookmarks"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:GemiNaut"
+        xmlns:local="clr-namespace:GemiNaut.Views"
         mc:Ignorable="d"
         Title="Bookmarks" Height="700" Width="1000">
     <Grid VerticalAlignment="Stretch">
