+++ GemiNaut/Views/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="GemiNaut.MainWindow"
+﻿<Window x:Class="GemiNaut.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:GemiNaut"
+        xmlns:local="clr-namespace:GemiNaut.Views"
         mc:Ignorable="d"
         Title="GemiNaut - a friendly Gemini browser" Height="600" Width="900"
         Closing="Window_Closing"
