+++ GemiNaut/Views/SettingsEditor.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="GemiNaut.SettingsEditor"
+﻿<Window x:Class="GemiNaut.Views.SettingsEditor"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:GemiNaut"
+        xmlns:local="clr-namespace:GemiNaut.Views"
         mc:Ignorable="d"
         WindowStartupLocation="CenterOwner"
         Title="Settings" Height="300" Width="700" MinHeight="400" MinWidth="700">
